+++ fw18-1070 YooxClone/src/App.jsx
@@ -1,25 +1,16 @@
 import { useState } from 'react'
 
 import './App.css'
+import HomePage from "./Sections/HomepageSection/HomePage"
 
-import HomepageMainPicture from './Sections/HompageMainPicture'
-import Navbar from './Sections/Navbar'
-import NewsletterSection from './Sections/NewsletterSection'
 
-import TwoPic from './Sections/TwoPic'
-
-import Footer from './Sections/Footer'
 
 function App() {
 
 
   return (
     <div className="App">
-      <Navbar/>
-      <HomepageMainPicture/>
-      <TwoPic/>
-      <NewsletterSection/>
-      <Footer/>
+      <HomePage/>
     </div>
   )
 }
