+++ fw18-1070 YooxClone/src/Sections/HomepageSection/Footer.jsx
@@ -1,6 +1,8 @@
 import { Box } from "@chakra-ui/react";
-import BottomPart from "../Components/Footer/BottomPart";
-import Social from "../Components/Footer/Social";
+import BottomPart from "../../Components/Footer/BottomPart"
+import Social from "../../Components/Footer/Social";
+
+
 
 export default function Footer(){
