+++ fw18-1070 YooxClone/src/Sections/HomepageSection/HompageMainPicture.jsx
@@ -1,4 +1,5 @@
-import BackgPic from "../Components/HomepageMainPicComps/BackgPic";
+import BackgPic from "../../Components/HomepageMainPicComps/BackgPic";
+
 
 export default function HomepageMainPicture(){
