+++ fw18-1070 YooxClone/src/Sections/HomepageSection/NewsletterSection.jsx
@@ -1,0 +1,10 @@
+import NewsLetter from "../../Components/NewsLetter/NewsLetter";
+
+
+
+export default function NewsletterSection(){
+
+    return(
+        <NewsLetter/>
+    )
+}
