+++ fw18-1070 YooxClone/src/Sections/HomepageSection/TwoPic.jsx
@@ -1,4 +1,5 @@
-import BfS from "../Components/BeforeFooter/BfS";
+import BfS from "../../Components/BeforeFooter/BfS";
+
 
 export default function TwoPic(){
