--- fw18-1070 YooxClone/src/Sections/NewsletterSection.jsx
@@ -1,9 +1,0 @@
-import BfS from "../Components/BeforeFooter/BfS";
-import NewsLetter from "../Components/NewsLetter/NewsLetter";
-
-export default function NewsletterSection(){
-
-    return(
-        <NewsLetter/>
-    )
-}
